+++ packages/playwright-chromium/install.js
@@ -1,39 +1,17 @@
-downloadBrowser('chromium', require('./index').chromium);
-
-async function downloadBrowser(browser, playwright) {
-  let progressBar = null;
-  let lastDownloadedBytes = 0;
-  function onProgress(downloadedBytes, totalBytes) {
-    if (!progressBar) {
-      const ProgressBar = require('progress');
-      progressBar = new ProgressBar(`Downloading ${browser} ${playwright._revision} - ${toMegabytes(totalBytes)} [:bar] :percent :etas `, {
-        complete: '=',
-        incomplete: ' ',
-        width: 20,
-        total: totalBytes,
-      });
-    }
-    const delta = downloadedBytes - lastDownloadedBytes;
-    lastDownloadedBytes = downloadedBytes;
-    progressBar.tick(delta);
-  }
-
-  const fetcher = playwright._createBrowserFetcher();
-  const revisionInfo = fetcher.revisionInfo();
-  await fetcher.download(revisionInfo.revision, onProgress);
-  logPolitely(`${browser} downloaded to ${revisionInfo.folderPath}`);
-}
-
-function toMegabytes(bytes) {
-  const mb = bytes / 1024 / 1024;
-  return `${Math.round(mb * 10) / 10} Mb`;
-}
-
-function logPolitely(toBeLogged) {
-  const logLevel = process.env.npm_config_loglevel;
-  const logLevelDisplay = ['silent', 'error', 'warn'].indexOf(logLevel) > -1;
-
-  if (!logLevelDisplay)
-    console.log(toBeLogged);
-}
-
+/**
+ * Copyright (c) Microsoft Corporation.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+const {downloadBrowser} = require('playwright-core/download-browser');
+downloadBrowser('chromium');
