+++ packages/playwright-firefox/package.json
@@ -1,7 +1,7 @@
 {
   "name": "playwright-firefox",
   "version": "0.9.17-post",
-  "description": "A high-level API to automate web browsers",
+  "description": "A high-level API to automate Firefox",
   "repository": "github:Microsoft/playwright",
   "main": "index.js",
   "scripts": {
