+++ packages/playwright-webkit/package.json
@@ -1,7 +1,7 @@
 {
   "name": "playwright-webkit",
   "version": "0.9.17-post",
-  "description": "A high-level API to automate web browsers",
+  "description": "A high-level API to automate WebKit",
   "repository": "github:Microsoft/playwright",
   "main": "index.js",
   "scripts": {
